Deleting files (images stored when saving/editing/deleting)

+++ controllers/admin.js
@@ -1,5 +1,6 @@
 const Product = require('../models/product');
 const {validationResult} = require('express-validator');
+const fileHelper = require('../util/file');
 
 exports.getAddProduct = (req, res, next) => {
     res.render('admin/edit-product',{
@@ -150,6 +151,7 @@ exports.postEditProduct = (req,res,next) => {
         product.price = updatedPrice;
         product.description = updatedDescription;
         if(image){
+            fileHelper.deleteFile(product.imageUrl);
             product.imageUrl = image.path;
         }
         
@@ -187,12 +189,18 @@ exports.getProducts = (req,res,next) => {
 
 exports.postDelete = (req,res,next) => {
     const prodId = req.body.productId;
-    Product.deleteOne({_id: prodId, userId: req.user._id})
-        .then(result => {
+    Product.findById(prodId).then(product => {
+        if(!product){
+            return next(new Error('Product not found'));
+        }
+        fileHelper.deleteFile(product.imageUrl);
+        return Product.deleteOne({_id: prodId, userId: req.user._id});
+    })
+    .then(result => {
             console.log('Product destroyed');
             res.redirect('/admin/products');
         })
-        .catch(err => {
+    .catch(err => {
             const error = new Error(err);
             error.httpStatusCode = 500;
             return next(error);
